pyproject.toml
[tool.poetry]
name = "slip-stream"
version = "0.1.0"
description = "JSON Schema-driven hexagonal backend framework for FastAPI + MongoDB"
authors = ["draw"]
license = "MIT"
readme = "README.md"
packages = [{include = "slip_stream"}]
keywords = ["fastapi", "mongodb", "hexagonal", "schema-driven", "crud", "pydantic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]

[tool.poetry.urls]
Homepage = "https://github.com/digitally-rendered/slip-stream"
Repository = "https://github.com/digitally-rendered/slip-stream"
Documentation = "https://github.com/digitally-rendered/slip-stream/tree/main/docs"
Changelog = "https://github.com/digitally-rendered/slip-stream/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/digitally-rendered/slip-stream/issues"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.115.0"
motor = "^3.4.0"
pydantic = ">=2.0.0"
pymongo = ">=4.0.0"
pyyaml = {version = ">=6.0", optional = true}
xmltodict = {version = ">=0.13.0", optional = true}
httpx = {version = ">=0.27.0", optional = true}
strawberry-graphql = {version = ">=0.235.0", optional = true, extras = ["fastapi"]}
mcp = {version = ">=1.0.0", optional = true}
sqlalchemy = {version = ">=2.0.0", optional = true, extras = ["asyncio"]}
schemathesis = {version = ">=3.32.0", optional = true}
opentelemetry-api = {version = ">=1.20.0", optional = true}
opentelemetry-sdk = {version = ">=1.20.0", optional = true}
dotted-dict = "^1.1.3"

[tool.poetry.extras]
yaml = ["pyyaml"]
xml = ["xmltodict"]
remote = ["httpx"]
graphql = ["strawberry-graphql"]
mcp = ["mcp"]
sql = ["sqlalchemy"]
test = ["schemathesis"]
telemetry = ["opentelemetry-api", "opentelemetry-sdk"]
all = ["pyyaml", "xmltodict", "httpx", "strawberry-graphql", "mcp", "sqlalchemy", "schemathesis", "opentelemetry-api", "opentelemetry-sdk"]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.24.0"
mongomock-motor = ">=0.0.34"
httpx = ">=0.27.0"
black = ">=24.0.0"
mypy = ">=1.0.0"
ruff = ">=0.4.0"
pyyaml = ">=6.0"
xmltodict = ">=0.13.0"
pdoc = "^16.0.0"
strawberry-graphql = {version = ">=0.235.0", extras = ["fastapi"]}
mcp = ">=1.0.0"
sqlalchemy = "^2.0.46"
aiosqlite = "^0.22.1"
schemathesis = ">=3.32.0"
opentelemetry-api = ">=1.20.0"
opentelemetry-sdk = ">=1.20.0"
greenlet = "^3.3.2"
pre-commit = "^4.5.1"
types-pyyaml = "^6.0.12.20250915"
pytest-cov = "^7.0.0"
pytest-benchmark = "^5.2.3"

[tool.poetry.scripts]
slip = "slip_stream.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E402",   # module-import-not-at-top (needed for conditional imports)
    "E501",   # line-too-long (black handles line length)
    "F541",   # f-string-missing-placeholders (false positives on template strings)
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"slip_stream/__init__.py" = ["F401"]  # public re-exports

[tool.ruff.lint.isort]
known-first-party = ["slip_stream"]

[tool.coverage.run]
source = ["slip_stream"]
omit = [
    "slip_stream/testing/*",
    "slip_stream/mcp/__main__.py",
]

[tool.coverage.report]
fail_under = 75
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
    "pass",
]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
exclude = [
    "examples/",
    "tests/",
]

